"Configuration Spring Security và xây dựng tính năng lưu trữ ảnh trên cloudinary và xây dựng tính năng updateProfile cho người dùng"

+++ src/main/java/vn/edu/fpt/repository/WardRepository.java
@@ -8,7 +8,7 @@ import vn.edu.fpt.model.Ward;
 import java.util.List;
 
 @Repository
-public interface WardRepository extends JpaRepository<Ward, Integer> {
+public interface WardRepository extends JpaRepository<Ward, Long> {
     Ward findByNameAndCity(String name, City city);
 
     List<Ward> findByCityId(Long cityId);
